Renamed error subfolders to use camelcase; moved error parent classes into their subfolders; added index files to subfolders.
