+++ app/errors/apiErrors/index.js
@@ -1,0 +1,7 @@
+// @flow
+
+export { default as Http5xxServerError } from './subclasses/Http5xxServerError';
+export { default as Http401UnauthorizedError } from './subclasses/Http401UnauthorizedError';
+export { default as Http403ForbiddenError } from './subclasses/Http403ForbiddenError';
+export { default as Http422ValidationError } from './subclasses/Http422ValidationError';
+export { default as UnexpectedHttpStatusError } from './subclasses/UnexpectedHttpStatusError';
