+++ app/errors/apiErrors/subclasses/UnexpectedHttpStatusError.js
