+++ app/errors/implementationErrors/subclasses/CorruptedInternalStateError.js
