+++ app/errors/implementationErrors/subclasses/UnsupportedOperationError.js
