+++ app/errors/index.js
@@ -1,14 +1,5 @@
 // @flow
 
-export { default as Http5xxServerError } from './api-errors/Http5xxServerError';
-export { default as Http401UnauthorizedError } from './api-errors/Http401UnauthorizedError';
-export { default as Http403ForbiddenError } from './api-errors/Http403ForbiddenError';
-export { default as Http422ValidationError } from './api-errors/Http422ValidationError';
-export { default as UnexpectedHttpStatusError } from './api-errors/UnexpectedHttpStatusError';
-
-export { default as CorruptedInternalStateError } from './implementation-errors/CorruptedInternalStateError';
-export { default as InvalidArgumentError } from './implementation-errors/InvalidArgumentError';
-export { default as NotYetImplementedError } from './implementation-errors/NotYetImplementedError';
-export { default as UnsupportedOperationError } from './implementation-errors/UnsupportedOperationError';
-
-export { default as ObjectNotFoundError } from './usage-errors/ObjectNotFoundError';
+export * from './apiErrors';
+export * from './implementationErrors';
+export * from './usageErrors';
