+++ app/errors/usageErrors/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as ObjectNotFoundError } from './subclasses/ObjectNotFoundError';
